temp_extracted/Efkv15-main/requirements.txt
python-dotenv
motor
pymongo
dnspython
paramiko
matplotlib
numpy
pandas
psutil
aiohttp
aiofiles
pytz
aiofiles
aiohttp
dnspython
matplotlib
motor
numpy
pandas
paramiko
psutil
pymongo
python-dotenv
pytz
aiofiles==24.1.0
aiohttp==3.11.0
asyncio==3.4.3
asyncssh==2.14.2
dnspython==2.7.0
motor==3.3.2
paramiko==3.5.1
pillow==10.2.0
psutil==5.9.0
pydantic==2.11.0
pymongo==4.6.1
python-dotenv==1.1.0
pytz==2024.1
requests==2.31.0
dnspython
motor
pymongo
python-dotenv
flask
aiofiles
aiohttp
asyncio
dnspython
matplotlib
motor
numpy
pandas
paramiko
psutil
pydantic
pymongo
python-dotenv
pytz
requests
aiofiles
aiohttp
asyncio
dnspython
flask
matplotlib
motor
numpy
pandas
paramiko
pillow
psutil
pydantic
pymongo
python-dotenv
pytz
requests
flask
flask
python-dotenv
py-cord==2.6.1
discord-py
py-cord==2.6.1
flask
psutil
dnspython
py-cord
python-dotenv
dnspython
py-cord
python-dotenv
dnspython
flask
py-cord
python-dotenv
